+++ backend/canisters/group_index/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Return newly created channel Ids after creating community ([#8156](https://github.com/open-chat-labs/open-chat/pull/8156))
+
 ## [[2.0.1784](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1784-group_index)] - 2025-06-10
 
 ### Changed